- revamp and 1.12.2 compatible - WalkToLocation,DeliverCookie,TalkToStrangers
