+++ PathfinderAPI/pom.xml
@@ -5,12 +5,12 @@
     <parent>
         <artifactId>PathfindergoalAPI</artifactId>
         <groupId>com.github.ysl3000.pathfindergoalapi</groupId>
-        <version>1.0-SNAPSHOT</version>
+        <version>1.3-SNAPSHOT</version>
     </parent>
-    <modelVersion>4.0.0</modelVersion>
 
     <artifactId>PathfinderAPI</artifactId>
-
+    <version>${parent.version}</version>
+    <modelVersion>4.0.0</modelVersion>
 
     <build>
         <finalName>${project.name}</finalName>
@@ -36,6 +36,14 @@
                 <configuration>
                     <show>private</show>
                     <nohelp>true</nohelp>
+                </configuration>
+            </plugin>
+            <plugin>
+                <groupId>org.apache.maven.plugins</groupId>
+                <artifactId>maven-compiler-plugin</artifactId>
+                <configuration>
+                    <source>8</source>
+                    <target>8</target>
                 </configuration>
             </plugin>
         </plugins>
