+++ PathfinderAPI/src/main/java/com/github/ysl3000/bukkit/pathfinding/pathfinding/ControllerJump.java
@@ -1,0 +1,16 @@
+package com.github.ysl3000.bukkit.pathfinding.pathfinding;
+
+/**
+ * Created by Yannick on 30.11.2016.
+ */
+public interface ControllerJump {
+    /**
+     * Lets the entity jump
+     */
+    void jump();
+
+    /**
+     * Sets whether the entity should jump all time (don't know what it does)
+     */
+    void setJumping();
+}
