+++ PathfinderAPI/src/main/java/com/github/ysl3000/bukkit/pathfinding/pathfinding/ControllerMove.java
@@ -1,0 +1,23 @@
+package com.github.ysl3000.bukkit.pathfinding.pathfinding;
+
+public interface ControllerMove {
+
+    boolean isOperationMove();
+
+    void move(double motionX, double motionY, double motionZ, double speed);
+
+    void move(float forward, float sideward);
+
+    void update();
+
+
+    double getX();
+
+    double getY();
+
+    double getZ();
+
+    double getSpeed();
+
+
+}
