--- PathfinderAPI/src/main/java/org/bukkit/pathfinding/ControllerMove.java
@@ -1,23 +1,0 @@
-package org.bukkit.pathfinding;
-
-public interface ControllerMove {
-
-    boolean isOperationMove();
-
-    void move(double motionX, double motionY, double motionZ, double speed);
-
-    void move(float forward, float sideward);
-
-    void update();
-
-
-    double getX();
-
-    double getY();
-
-    double getZ();
-
-    double getSpeed();
-
-
-}
