+++ PathfinderPlugin/pom.xml
@@ -5,11 +5,13 @@
     <parent>
         <artifactId>PathfindergoalAPI</artifactId>
         <groupId>com.github.ysl3000.pathfindergoalapi</groupId>
-        <version>1.0-SNAPSHOT</version>
+        <version>1.3-SNAPSHOT</version>
     </parent>
     <modelVersion>4.0.0</modelVersion>
 
     <artifactId>PathfinderPlugin</artifactId>
+    <version>${parent.version}</version>
+
     <build>
         <plugins>
             <plugin>
@@ -21,8 +23,6 @@
                     <descriptorRefs>
                         <descriptorRef>jar-with-dependencies</descriptorRef>
                     </descriptorRefs>
-                    <finalName>${project.name}</finalName>
-                    <appendAssemblyId>false</appendAssemblyId>
                 </configuration>
                 <executions>
                     <execution>
@@ -44,7 +44,7 @@
         <dependency>
             <groupId>com.github.ysl3000.pathfindergoalapi</groupId>
             <artifactId>PathfinderAPI</artifactId>
-            <version>1.0-SNAPSHOT</version>
+            <version>${parent.version}</version>
             <scope>compile</scope>
         </dependency>
