--- PathfinderPlugin/src/main/java/org/bukkit/craftbukkit/v1_12_R1/PathfinderGoalPlugin.java
@@ -1,25 +1,0 @@
-package org.bukkit.craftbukkit.v1_12_R1;
-
-import org.bukkit.PathfinderGoalAPI;
-import org.bukkit.craftbukkit.v1_12_R1.pathfinding.CraftPathfinderManager;
-import org.bukkit.plugin.java.JavaPlugin;
-
-public class PathfinderGoalPlugin extends JavaPlugin {
-
-    private CraftPathfinderManager craftPathfinderManager;
-
-    @Override
-    public void onDisable() {
-
-        this.craftPathfinderManager=null;
-
-    }
-
-    @Override
-    public void onEnable() {
-
-        this.craftPathfinderManager = new CraftPathfinderManager();
-        PathfinderGoalAPI.setApi(this.craftPathfinderManager);
-
-    }
-}
