+++ PathfinderTestPlugin/src/main/java/com/github/ysl3000/bukkit/test/ICommand.java
@@ -1,0 +1,12 @@
+package com.github.ysl3000.bukkit.test;
+
+import org.bukkit.entity.Player;
+
+import java.util.List;
+
+/**
+ * Created by ysl3000
+ */
+public interface ICommand {
+    boolean execute(Player p, List<String> args);
+}
