+++ pom.xml
@@ -7,26 +7,17 @@
     <groupId>com.github.ysl3000.pathfindergoalapi</groupId>
     <artifactId>PathfindergoalAPI</artifactId>
     <packaging>pom</packaging>
-    <version>1.0-SNAPSHOT</version>
+    <version>1.3-SNAPSHOT</version>
     <modules>
         <module>PathfinderAPI</module>
+        <module>PathfinderPlugin</module>
+        <module>PathfinderTestPlugin</module>
     </modules>
 
     <properties>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
-        <spigot.version>1.12-R0.1-SNAPSHOT</spigot.version>
+        <spigot.version>1.12.2-R0.1-SNAPSHOT</spigot.version>
         <maven.compiler.source>1.8</maven.compiler.source>
         <maven.compiler.target>1.8</maven.compiler.target>
     </properties>
-
-    <profiles>
-        <profile>
-            <id>Impl</id>
-            <modules>
-                <module>PathfinderPlugin</module>
-            </modules>
-        </profile>
-    </profiles>
-
-
 </project>
